Разделяет утилитарный модуль на общий и зависимости для фильмов
