+++ src/utils/film.js
@@ -1,0 +1,13 @@
+import dayjs from 'dayjs';
+
+const formatDate = (date) => dayjs(date);
+
+const getShortDescription = (description) => {
+  if (description.trim().length > 140) {
+    return `${description.trim().slice(0, 139)}...`;
+  }
+
+  return description.trim();
+};
+
+export {formatDate, getShortDescription};
